backend/requirements.txt
# FastAPI e dependências
fastapi==0.115.0
uvicorn==0.27.0
starlette==0.38.6

# MongoDB (usando PyMongo como alternativa ao Motor)
pymongo==4.6.0

# Pydantic (para validação de dados)
pydantic==2.10.6

# Variáveis de ambiente
python-dotenv==0.19.0

# Mercado Pago SDK
mercadopago==2.2.0

# Dependências adicionais comuns em projetos FastAPI
httpx==0.27.0  # Cliente HTTP assíncrono
python-multipart==0.0.7  # Para upload de arquivos
Jinja2==3.1.4  # Para templates HTML (se necessário)
